Adapt TestLogger class to \Psr\Log\AbstractLogger version in /vendor
Depending on version of php \Psr\Log\AbstractLogger::log() method
may or may not declare a return type of void.

Detect signature of AbstractLogger and include TestLogger class
whose signature mirrors the class it extends.
